requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altgraph==0.17.5
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.11.2
attrs==25.4.0
beautifulsoup4==4.14.3
bs4==0.0.2
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
cryptography==46.0.5
curl_cffi==0.13.0
distro==1.9.0
et_xmlfile==2.0.0
frozendict==2.4.7
frozenlist==1.8.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.190.0
google-auth==2.49.0.dev0
google-auth-httplib2==0.3.0
google-genai==1.64.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
grpcio==1.78.1
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
multidict==6.7.1
multitasking==0.0.12
numpy==2.4.2
openpyxl==3.1.5
packaging==26.0
pandas==3.0.1
peewee==4.0.0
pefile==2024.8.26
pillow==12.1.1
platformdirs==4.9.2
propcache==0.4.1
proto-plus==1.27.1
protobuf==5.29.6
psutil==7.2.2
pyaes==1.6.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyinstaller==6.19.0
pyinstaller-hooks-contrib==2026.1
pyparsing==3.3.2
pystray==0.19.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-telegram-bot==22.6
pytz==2025.2
pywin32-ctypes==0.2.3
requests==2.32.5
rsa==4.9.1
setuptools==82.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
Telethon==1.42.0
tenacity==9.1.4
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
websockets==15.0.1
yarl==1.22.0
yfinance==1.2.0
